requirements.txt
Flask
Flask-CORS
unidecode
pymupdf
spacy
pymongo
PyPDF2
python-dotenv
openai
azure-ai-documentintelligence
azure-core
sentence-transformers
numpy
gunicorn
